// src/app/pages/common/photo-gallery/shared/photo-gallery-enum.ts
export enum PhotoGalleryEnum {
    galleryTitle = 'Photo Gallery',
    dragAndDropText = 'Drag and drop your photos here',
    uploadPhotosTitle ='Upload at least 1 photo',
    uploadPhotosSubTitle ="You'll also be able to upload more after registration",
    sectionTips ='No professional photos? No problem.',
    useDevices = 'You can use: ',
    useSmartphone = ' A smartphone ',
    useDigitalCamera = ' A digital camera',
    nameTip = 'Tip! ', 
    linkTips ='Guests love photos! Here are some tips to help you take great photos of your property:',
    copyrightPhotos ="If you don't know who made a photo, it's best to avoid using it. Only use photos that you own the copyright of yourself, or if taken by someone else, make sure you have the photographer's consent to use the photo.",

    dragarea = 'dragarea',
    droparea = 'droparea',
    previewElement = 'preview',
    spanElement = 'span',
    inputElement = 'input',
    typeElement = 'checkbox',
    idAttribute = 'id',
    checkboxId = 'check',
    typeAttribute = 'type',
    typeFile = 'image/',
    tagImage = 'img',
    fileAttribute = 'file',
    imageClass = "images-style",
    spanClass = 'span-checkbox-images',
    checkboxClass = 'chechbox-images',
    listenerClick = 'click',
    componentForm = 'photos',
    incompleteString = 'incomplete',
    dragulaItem = 'image-item',

    dragover = 'dragover',
    dragenter = 'dragenter',
    dragend = 'dragend',
    dragleave = 'dragleave',
    drop = 'drop',

    valueZero = 0,
    valueOne = 1,
    postMethod = 'POST',
    putMethod = 'PUT',
    deleteMethod = 'DELETE',
    negativeValue = -1,

    deletePhotosMessage = "Are you sure you want to delete these photos?",
}